+++ packages/w3p/src/types/solana.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   PublicKey,
   SendOptions,
   Transaction,